+++ qcs-api-client-openapi/internal/src/models/boxcar_averager_readout_kernel.rs
@@ -16,28 +16,28 @@ pub struct BoxcarAveragerReadoutKernel {
     #[serde(rename = "_type", skip_serializing_if = "Option::is_none")]
     pub _type: Option<Type>,
     #[serde(rename = "bias")]
-    pub bias: f32,
+    pub bias: f64,
     #[serde(rename = "detuning")]
-    pub detuning: f32,
+    pub detuning: f64,
     #[serde(rename = "duration")]
-    pub duration: f32,
+    pub duration: f64,
     #[serde(rename = "frame")]
     pub frame: String,
     #[serde(rename = "phase")]
-    pub phase: f32,
+    pub phase: f64,
     #[serde(rename = "scale")]
-    pub scale: f32,
+    pub scale: f64,
 }
 
 impl BoxcarAveragerReadoutKernel {
     /// Readout kernel that is a constant, with a normalization condition.  The normalization rescales the values such that the kernel integrates to 1, ie. makes the scale equal to 1/duration. This is useful for preserving information about the amplitude of the readout pulse that is demodulated, however this message puts the normalization on the sequencer where it suffers from discretization error and actually goes to zero amplitude at a duration of >16 microseconds.  See also https://gitlab.com/rigetti/qcs/pidgin/-/issues/328.
     pub fn new(
-        bias: f32,
-        detuning: f32,
-        duration: f32,
+        bias: f64,
+        detuning: f64,
+        duration: f64,
         frame: String,
-        phase: f32,
-        scale: f32,
+        phase: f64,
+        scale: f64,
     ) -> BoxcarAveragerReadoutKernel {
         BoxcarAveragerReadoutKernel {
             _type: None,
